refactor: Message 위치 이동
Message가 저장되어 있는 enum을 config 패키지로 이동
