+++ src/main/java/lotto/config/Information.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.config;
 
 public enum Information {
     INPUT_PURCHASE_AMOUNT("구입금액을 입력해 주세요."),
@@ -6,9 +6,9 @@ public enum Information {
     INPUT_WINNING_NUMBER("당첨 번호를 입력해 주세요."),
     INPUT_BONUS_NUMBER("보너스 번호를 입력해 주세요."),
     INPUT_WINNING_STATISTICS("당첨 통계"),
+    CONTOUR("---"),
     RESULT_WINNING_BASIC_FORMAT("%d개 일치 (%d원) - %d개"),
     RESULT_WINNING_SECOND_FORMAT("%d개 일치, 보너스 볼 일치 (%d원) - %d개"),
-    CONTOUR("---"),
     NEW_LINE("\n");
 
     private final String information;
