+++ src/main/java/lotto/config/LottoConstants.java
@@ -1,0 +1,11 @@
+package lotto.config;
+
+public final class LottoConstants {
+    public static final int MIN_RANGE = 1;
+    public static final int MAX_RANGE = 45;
+    public static final int LOTTO_NUMBER_COUNT = 6;
+    public static final int BONUS_NUMBER_SIZE = 1;
+
+    private LottoConstants() {
+    }
+}
